routing: recover proper error position reporting
This commit fixes the test that the previous commit broke. This
approach is less elegant but it preserves the current error structure.
If we are willing to abandon position based errors, and go to vertex
based ones we could simply collect them using the function defined
in the prior commit. There are other ways to preserve position
reporting, but I went with this way for the time being to ensure the
tests would pass without changes.